finished up skeeball, or at least as far as it is going to go

+++ VR23/Assets/HoleTrigger.cs
@@ -7,7 +7,7 @@ public class HoleTrigger : MonoBehaviour
     [SerializeField]
     Skeeball machine;
 
-    public float pointValue;
+    public int pointValue;
     // Start is called before the first frame update
     void Start()
     {
